try to fix task
